+++ src/greentest/monkey_test.py
@@ -37,5 +37,5 @@ else:
 with module_file:
     module_source = module_file.read()
 module_source = disable_tests_in_source(module_source, test_name)
-module_code = compile(module_source, test_filename, 'exec')
+module_code = compile(module_source, os.path.abspath(test_filename), 'exec')
 exec(module_code, globals())